# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API de gestión de compañías y empleados",
        "title": "Compania API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/api/empleados": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "empleados"
                ],
                "summary": "Obtener todos los empleados",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "empleados"
                ],
                "summary": "Crear un empleado",
                "parameters": [
                    {
                        "description": "Datos del empleado",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateEmpleadoRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/empleados/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "empleados"
                ],
                "summary": "Obtener un empleado por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del empleado",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "empleados"
                ],
                "summary": "Actualizar un empleado",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del empleado",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos del empleado",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UpdateEmpleadoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "empleados"
                ],
                "summary": "Eliminar un empleado",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del empleado",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dtos.CreateEmpleadoRequest": {
            "type": "object",
            "required": [
                "apellido",
                "cargo",
                "compania_id",
                "correo",
                "nombre",
                "salario"
            ],
            "properties": {
                "apellido": {
                    "type": "string"
                },
                "cargo": {
                    "type": "string"
                },
                "compania_id": {
                    "type": "integer"
                },
                "correo": {
                    "type": "string"
                },
                "nombre": {
                    "type": "string"
                },
                "salario": {
                    "type": "number"
                }
            }
        },
...
        "dtos.UpdateEmpleadoRequest": {
            "type": "object",
            "required": [
                "apellido",
                "cargo",
                "correo",
                "nombre",
                "salario"
            ],
            "properties": {
                "apellido": {
                    "type": "string"
                },
                "cargo": {
                    "type": "string"
                },
                "correo": {
                    "type": "string"
                },
                "nombre": {
                    "type": "string"
                },
                "salario": {
                    "type": "number"
                }
            }
        }
    }
}